fix(sdk): Move set/remove tag methods to room::Common
