+++ crates/matrix-sdk/src/room/common.rs
@@ -7,9 +7,11 @@ use ruma::{
         membership::{get_member_events, join_room_by_id, leave_room},
         message::get_message_events,
         room::get_room_event,
+        tag::{create_tag, delete_tag},
     },
     events::{
-        room::history_visibility::HistoryVisibility, AnyStateEvent, AnySyncStateEvent, EventType,
+        room::history_visibility::HistoryVisibility, tag::TagInfo, AnyStateEvent,
+        AnySyncStateEvent, EventType,
     },
     serde::Raw,
     EventId, UserId,
@@ -18,7 +20,7 @@ use ruma::{
 use crate::{
     media::{MediaFormat, MediaRequest, MediaType},
     room::RoomType,
-    BaseRoom, Client, Result, RoomMember,
+    BaseRoom, Client, HttpError, HttpResult, Result, RoomMember,
 };
 
 /// A struct containing methods that are common for Joined, Invited and Left
@@ -443,5 +445,51 @@ impl Common {
         }
 
         Ok(true)
+    }
+
+    /// Adds a tag to the room, or updates it if it already exists.
+    ///
+    /// Returns the [`create_tag::Response`] from the server.
+    ///
+    /// # Arguments
+    /// * `tag` - The tag to add or update.
+    ///
+    /// * `tag_info` - Information about the tag, generally containing the
+    ///   `order` parameter.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// # use ruma::events::tag::TagInfo;
+    /// # futures::executor::block_on(async {
+    /// # let homeserver = url::Url::parse("http://localhost:8080")?;
+    /// # let mut client = matrix_sdk::Client::new(homeserver)?;
+    /// # let room_id = matrix_sdk::ruma::room_id!("!test:localhost");
+    /// use matrix_sdk::ruma::events::tag::TagInfo;
+    ///
+    /// if let Some(room) = client.get_joined_room(&room_id) {
+    ///     let mut tag_info = TagInfo::new();
+    ///     tag_info.order = Some(0.9);
+    ///
+    ///     room.set_tag("u.work", tag_info ).await?;
+    /// }
+    /// # Result::<_, matrix_sdk::Error>::Ok(()) });
+    /// ```
+    pub async fn set_tag(&self, tag: &str, tag_info: TagInfo) -> HttpResult<create_tag::Response> {
+        let user_id = self.client.user_id().await.ok_or(HttpError::AuthenticationRequired)?;
+        let request = create_tag::Request::new(&user_id, self.inner.room_id(), tag, tag_info);
+        self.client.send(request, None).await
+    }
+
+    /// Removes a tag from the room.
+    ///
+    /// Returns the [`delete_tag::Response`] from the server.
+    ///
+    /// # Arguments
+    /// * `tag` - The tag to remove.
+    pub async fn remove_tag(&self, tag: &str) -> HttpResult<delete_tag::Response> {
+        let user_id = self.client.user_id().await.ok_or(HttpError::AuthenticationRequired)?;
+        let request = delete_tag::Request::new(&user_id, self.inner.room_id(), tag);
+        self.client.send(request, None).await
     }
 }
